* develop: (395 commits)
  set default server address to 0.0.0.0 to use pf from other machines than localhost
  Change version to 6.15.0-SNAPSHOT
  Change version to 6.14, Edit NEWS.md
  PROJECTFORGE-3209 fix admin system page is now only viewable for admins
  PROJECTFORGE-1724: - update luceane info page link
  Fix typo in localization
  fix typo in i18n file
  PROJECTFORGE-2920 fix security issue
  PROJECTFORGE-3179 Fix NPE, change some code for line return
  PROJECTFORGE-3179 Fix line return and filename of generted invoice
  PROJECTFORGE-3176 adapt excel import replacement to new excel package to make the groovy scripts great again
  PROJECTFORGE-3192 fix bug with liquidity forecast
  PROJECTFORGE-3177: Fix bug in export, only names will be displayed
  PROJECTFORGE-3172: Fix bug in copy timesheets PROJECTFORGE-3174: Fix bug in access handling
  PROJECTFORGE-3190 Fix unused vacation from previous year bug
  PROJECTFORGE-3178 fix excel export problem with more than 25 columns and remove the limit of 27 filter columns
  PROJECTFORGE-3173 require date and amount in payment schedule, add missing labels for required error message, fix findbugs issue
  PROJECTFORGE-3166 validate dates in payment schedule
  Change version to 6.14.0-SNAPSHOT
  Change version to 6.13.0, change NEWS.md
  ...

# Conflicts:
#	projectforge-application/application.properties

+++ projectforge-wicket/src/main/java/org/projectforge/web/vacation/VacationEditPage.java
@@ -27,11 +27,12 @@ import org.apache.log4j.Logger;
 import org.apache.wicket.markup.html.WebPage;
 import org.apache.wicket.request.mapper.parameter.PageParameters;
 import org.apache.wicket.spring.injection.annot.SpringBean;
+import org.projectforge.business.configuration.ConfigurationService;
 import org.projectforge.business.fibu.api.EmployeeService;
-import org.projectforge.business.user.I18nHelper;
 import org.projectforge.business.vacation.model.VacationDO;
 import org.projectforge.business.vacation.model.VacationStatus;
 import org.projectforge.business.vacation.service.VacationService;
+import org.projectforge.framework.i18n.I18nHelper;
 import org.projectforge.web.fibu.ISelectCallerPage;
 import org.projectforge.web.wicket.AbstractEditPage;
 import org.projectforge.web.wicket.AbstractSecuredBasePage;
@@ -47,6 +48,9 @@ public class VacationEditPage extends AbstractEditPage<VacationDO, VacationEditF
 
   @SpringBean
   private VacationService vacationService;
+
+  @SpringBean
+  private ConfigurationService configService;
 
   @SpringBean
   private EmployeeService employeeService;
@@ -129,17 +133,45 @@ public class VacationEditPage extends AbstractEditPage<VacationDO, VacationEditF
   public AbstractSecuredBasePage afterSaveOrUpdate()
   {
     try {
+      vacationService.saveOrUpdateVacationCalendars(form.getData(), form.assignCalendarListHelper.getAssignedItems());
       if (wasNew) {
         vacationService.sendMailToVacationInvolved(form.getData(), true, false);
-      } else {
-        if (VacationStatus.IN_PROGRESS.equals(form.getData().getStatus())) {
-          vacationService.sendMailToVacationInvolved(form.getData(), false, false);
-        }
+      } else if (VacationStatus.IN_PROGRESS == form.getData().getStatus()) {
+        vacationService.sendMailToVacationInvolved(form.getData(), false, false);
+      }
+      if (VacationStatus.APPROVED.equals(form.getData().getStatus())) {
+        //To intercept special cases for add or delete team calendars
+        vacationService.markTeamEventsOfVacationAsDeleted(form.getData(), false);
+        vacationService.createEventsForVacationCalendars(form.getData());
       }
       if (form.getStatusBeforeModification() != null) {
-        if (form.getStatusBeforeModification().equals(VacationStatus.IN_PROGRESS) && VacationStatus.APPROVED.equals(form.getData().getStatus())) {
-          vacationService.updateUsedVacationDaysFromLastYear(form.getData());
-          vacationService.sendMailToEmployeeAndHR(form.getData(), true);
+        if (form.getStatusBeforeModification() == VacationStatus.IN_PROGRESS) {
+          switch (form.getData().getStatus()) {
+            case APPROVED:
+              // IN_PROGRESS -> APPROVED
+              vacationService.updateUsedVacationDaysFromLastYear(form.getData());
+              vacationService.sendMailToEmployeeAndHR(form.getData(), true);
+              break;
+
+            case REJECTED:
+              // IN_PROGRESS -> REJECTED
+              vacationService.sendMailToEmployeeAndHR(form.getData(), false);
+              break;
+
+            default:
+              // nothing to do
+          }
+        }
+        if (form.getStatusBeforeModification() == VacationStatus.APPROVED) {
+          switch (form.getData().getStatus()) {
+            case REJECTED:
+            case IN_PROGRESS:  // APPROVED -> NOT APPROVED
+              vacationService.markTeamEventsOfVacationAsDeleted(form.getData(), true);
+              vacationService.deleteUsedVacationDaysFromLastYear(form.getData());
+              break;
+            default:
+              // nothing to do
+          }
         }
       }
     } catch (final Exception e) {
@@ -153,8 +185,11 @@ public class VacationEditPage extends AbstractEditPage<VacationDO, VacationEditF
   public WebPage afterDelete()
   {
     try {
-      vacationService.deleteUsedVacationDaysFromLastYear(form.getData());
-      vacationService.sendMailToVacationInvolved(form.getData(), false, true);
+      if (VacationStatus.APPROVED.equals(form.getData().getStatus())) {
+        vacationService.markTeamEventsOfVacationAsDeleted(form.getData(), true);
+        vacationService.deleteUsedVacationDaysFromLastYear(form.getData());
+        vacationService.sendMailToVacationInvolved(form.getData(), false, true);
+      }
     } catch (final Exception e) {
       log.error("There is a exception in afterDelete: " + e.getMessage(), e);
       error(I18nHelper.getLocalizedMessage("vacation.error.sendmail"));
@@ -162,11 +197,19 @@ public class VacationEditPage extends AbstractEditPage<VacationDO, VacationEditF
     return null;
   }
 
+  @Override
   public WebPage afterUndelete()
   {
     try {
-      vacationService.updateUsedVacationDaysFromLastYear(form.getData());
-      vacationService.sendMailToVacationInvolved(form.getData(), false, false);
+      vacationService.undeleteTeamEventsOfVacation(form.getData());
+      if (VacationStatus.APPROVED.equals(form.getData().getStatus())) {
+        vacationService.markAsUnDeleteEventsForVacationCalendars(form.getData());
+        vacationService.updateUsedVacationDaysFromLastYear(form.getData());
+        vacationService.sendMailToEmployeeAndHR(form.getData(), true);
+        vacationService.createEventsForVacationCalendars(form.getData());
+      } else {
+        vacationService.sendMailToVacationInvolved(form.getData(), false, false);
+      }
     } catch (final Exception e) {
       log.error("There is a exception in afterUndelete: " + e.getMessage(), e);
       error(I18nHelper.getLocalizedMessage("vacation.error.sendmail"));
